+++ packages/dart_firebase_admin/test/messaging/messaging_test.dart
@@ -1,10 +1,14 @@
+import 'dart:convert';
+
 import 'package:dart_firebase_admin/src/messaging.dart';
 import 'package:firebaseapis/fcm/v1.dart' as fmc1;
 import 'package:firebaseapis/fcm/v1.dart';
+import 'package:http/http.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 import '../google_cloud_firestore/util/helpers.dart';
+import '../mock.dart';
 
 class ProjectsMessagesResourceMock extends Mock
     implements ProjectsMessagesResource {}
@@ -17,8 +21,6 @@ class FirebaseCloudMessagingApiMock extends Mock
 
 class ProjectsResourceMock extends Mock implements ProjectsResource {}
 
-class SendMessageRequestFake extends Fake implements SendMessageRequest {}
-
 void main() {
   late Messaging messaging;
 
@@ -27,9 +29,7 @@ void main() {
   final projectResourceMock = ProjectsResourceMock();
   final messagingApiMock = FirebaseCloudMessagingApiMock();
 
-  setUpAll(() {
-    registerFallbackValue(SendMessageRequestFake());
-  });
+  setUpAll(registerFallbacks);
 
   void mockV1<T>() {
     when(() => requestHandler.v1<T>(any())).thenAnswer((invocation) async {
@@ -54,6 +54,73 @@ void main() {
     reset(messages);
     reset(projectResourceMock);
     reset(messagingApiMock);
+  });
+
+  group('Error handling', () {
+    for (final (:code, :error) in [
+      (code: 400, error: MessagingClientErrorCode.invalidArgument),
+      (code: 401, error: MessagingClientErrorCode.authenticationError),
+      (code: 403, error: MessagingClientErrorCode.authenticationError),
+      (code: 500, error: MessagingClientErrorCode.internalError),
+      (code: 503, error: MessagingClientErrorCode.serverUnavailable),
+      (code: 505, error: MessagingClientErrorCode.unknownError),
+    ]) {
+      test('converts $code codes into errors', () async {
+        final clientMock = ClientMock();
+        when(() => clientMock.send(any())).thenAnswer(
+          (_) => Future.value(
+            StreamedResponse(Stream.value(utf8.encode('')), code),
+          ),
+        );
+
+        final app = createApp(client: clientMock);
+        final handler = Messaging(app);
+
+        await expectLater(
+          () => handler.send(TokenMessage(token: '123')),
+          throwsA(
+            isA<FirebaseMessagingAdminException>()
+                .having((e) => e.errorCode, 'errorCode', error),
+          ),
+        );
+      });
+    }
+
+    for (final MapEntry(key: messagingError, value: code)
+        in messagingServerToClientCode.entries) {
+      test('converts $messagingError error codes', () async {
+        final clientMock = ClientMock();
+        when(() => clientMock.send(any())).thenAnswer(
+          (_) => Future.value(
+            StreamedResponse(
+              Stream.value(
+                utf8.encode(
+                  jsonEncode({
+                    'error': {'message': messagingError},
+                  }),
+                ),
+              ),
+              400,
+              headers: {
+                'content-type': 'application/json',
+              },
+            ),
+          ),
+        );
+
+        final app = createApp(client: clientMock);
+        final handler = Messaging(app);
+
+        await expectLater(
+          () => handler.send(TokenMessage(token: '123')),
+          throwsA(
+            isA<FirebaseMessagingAdminException>()
+                .having((e) => e.errorCode, 'errorCode', code)
+                .having((e) => e.code, 'code', 'messaging/${code.code}'),
+          ),
+        );
+      });
+    }
   });
 
   group('Messaging.send', () {
@@ -90,6 +157,11 @@ void main() {
         () => messaging.send(TopicMessage(topic: 'test')),
         throwsA(
           isA<FirebaseMessagingAdminException>()
+              .having(
+                (e) => e.errorCode,
+                'errorCode',
+                MessagingClientErrorCode.internalError,
+              )
               .having((e) => e.message, 'message', 'No name in response'),
         ),
       );